requirements.txt
# Core
fastapi==0.115.6
uvicorn[standard]==0.34.0
pydantic==2.10.4
pydantic-settings==2.7.1
python-dotenv==1.0.1
python-multipart==0.0.20

# Database — Supabase
supabase==2.11.0

# Transcription — Deepgram
deepgram-sdk==3.8.0

# LLM — Event extraction & Intent classification
openai==1.58.1
tiktoken==0.8.0

# Async HTTP
httpx==0.28.1

# Logging
structlog==24.4.0

# Testing
pytest==8.3.4
pytest-asyncio==0.25.0
